kubernetes: Add formatting on cluster displayed in prompt
Having the cluster name in the prompt is great. However, without
formatting, people can still mistake prod clusters with dev ones.
Especially when their names are pretty close (dev.fqdn.com,
prod.fqdn.com).

To prevent this, we apply a formatting based on the name of the cluster.
Any cluster containing dev in its name is considered "dev" which is
safe. Any other cluster are production ones and are highlighted in red.